Dev

See merge request s07-webmobile2-sub2/S07P12A506!278

+++ back/Trippy/src/main/java/com/ssafy/trippy/Controller/PostController.java
@@ -3,12 +3,13 @@ package com.ssafy.trippy.Controller;
 import com.ssafy.trippy.Domain.Location;
 import com.ssafy.trippy.Domain.Member;
 import com.ssafy.trippy.Dto.Request.RequestPostDto;
+import com.ssafy.trippy.Dto.Response.ResponseBadgeDto;
 import com.ssafy.trippy.Dto.Response.ResponsePostDto;
+import com.ssafy.trippy.Dto.Response.ResponseSavepostDto;
 import com.ssafy.trippy.Dto.Update.UpdatePostDto;
 import com.ssafy.trippy.Service.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.hibernate.sql.Update;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -16,7 +17,6 @@ import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -27,6 +27,8 @@ public class PostController {
     private final PostService postService;
 
     private final MemberService memberService;
+
+    private final BadgeService badgeService;
     private static final String SUCCESS = "OK";
     private static final String FAIL = "ERROR";
 
@@ -38,18 +40,31 @@ public class PostController {
         requestPostDto.setMember_id(memberId);
         try {
             Long id = postService.savePost(requestPostDto, images);
-            return new ResponseEntity<>(id, HttpStatus.OK);
+            ResponseSavepostDto responseSavepostDto = new ResponseSavepostDto(id);
+            Long cnt = postService.cntPostsByMemberId(memberId);
+            if(cnt==1L){
+                ResponseBadgeDto responseBadgeDto = badgeService.saveBadge(2L,memberId);
+                responseSavepostDto.addBadge(responseBadgeDto);
+            }
+            return new ResponseEntity<>(responseSavepostDto, HttpStatus.OK);
         } catch (Exception e) {
             e.printStackTrace();
             return new ResponseEntity<>("저장할 수 없습니다.", HttpStatus.BAD_REQUEST);
         }
     }
 
+
     @DeleteMapping("/auth/posts/{post_id}")
-    public ResponseEntity<?> deletePost(@PathVariable("post_id") Long post_id) {
+    public ResponseEntity<?> deletePost(HttpServletRequest request, @PathVariable("post_id") Long post_id) {
+        ResponsePostDto post = postService.findPostId(post_id);
+        Long memberId = memberService.getIdByToken(request.getHeader("X-AUTH-TOKEN"));
         try {
-            postService.deletePost(post_id);
-            return new ResponseEntity<>(SUCCESS, HttpStatus.OK);
+            if (memberId == post.getMemberId()) {
+                postService.deletePost(post_id);
+                return new ResponseEntity<>(SUCCESS, HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("본인이 작성한 글만 삭제가 가능합니다.", HttpStatus.BAD_REQUEST);
+            }
 
         } catch (Exception e) {
             e.printStackTrace();
@@ -58,23 +73,36 @@ public class PostController {
     }
 
     @PutMapping("/auth/posts/{post_id}")
-    public ResponseEntity<?> updatePost(@PathVariable("post_id") Long post_id, @RequestBody @Valid RequestPostDto requestPostDto) {
+    public ResponseEntity<?> updatePost(HttpServletRequest request, @PathVariable("post_id") Long
+            post_id, @RequestPart("post") @Valid RequestPostDto requestPostDto
+            , @RequestPart("images") List<MultipartFile> images) {
+        ResponsePostDto post = postService.findPostId(post_id);
+        Long memberId = memberService.getIdByToken(request.getHeader("X-AUTH-TOKEN"));
         try {
-            postService.updatePost(post_id, requestPostDto);
-            return new ResponseEntity<>(SUCCESS, HttpStatus.OK);
-
+            if (memberId == post.getMemberId()) {
+                Long id = postService.updatePost(post_id, requestPostDto, images);
+                return new ResponseEntity<>(id, HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("본인이 작성한 글만 수정이 가능합니다.", HttpStatus.BAD_REQUEST);
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return new ResponseEntity<>("수정할 수 없습니다.", HttpStatus.BAD_REQUEST);
         }
     }
 
+
     @GetMapping("/posts")
     public ResponseEntity<?> getAllPost() {
-        List<ResponsePostDto> responsePostDtos = new ArrayList<>();
+        List<ResponsePostDto> responsePostDtos;
         try {
             responsePostDtos = postService.findAll();
-            return new ResponseEntity<>(responsePostDtos, HttpStatus.OK);
+            if (responsePostDtos.size() == 0) {
+                return new ResponseEntity<>("작성한 글이 없습니다.", HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(responsePostDtos, HttpStatus.OK);
+
+            }
 
         } catch (Exception e) {
             e.printStackTrace();
@@ -100,7 +128,11 @@ public class PostController {
         Long memberId = memberService.getIdByToken(request.getHeader("X-AUTH-TOKEN"));
         try {
             List<ResponsePostDto> responsePostDtos = postService.findAllByMember(Member.builder().id(memberId).build());
-            return new ResponseEntity<>(responsePostDtos, HttpStatus.OK);
+            if (responsePostDtos.size() == 0) {
+                return new ResponseEntity<>("작성한 게시글이 없습니다.", HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(responsePostDtos, HttpStatus.OK);
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return new ResponseEntity<>("게시물이 없습니다.", HttpStatus.BAD_REQUEST);
